Limit logistics arriving at beacon to 100 of one type
If sending multiple types of item to a single beacon, it could get blocked if the beacon gets full of one item. To prevent this, I limited the number of items of each type to 100. I wanted to limit it to a full stack, but I'm not sure how to check the size of a full stack of each type.